api/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
sqlparse==0.5.0
psycopg2
psycopg2-binary
git+https://github.com/dpkp/kafka-python.git
